Adding React Routing package to navigate to a separate detail page
- creating separate paths for recipes and recipe details
