+++ src/components/card/card.tsx
@@ -1,13 +1,9 @@
 import classes from './card.module.css';
 
 const Card = (props: any) => {
-  const clickHandler = () => {
-    props.clickHander(props.recipe)
-  };
-
   return (
-    <div className={classes.card} onClick={clickHandler}>
-      <p>{props.recipe.name}</p>
+    <div className={classes.card}>
+      <p>{props.name}</p>
     </div>
   );
 };
