+++ src/components/layout/layout.tsx
@@ -1,10 +1,9 @@
-import Recipes from '../recipes/recipes';
 import classes from './layout.module.css';
 
-const Layout = () => {
+const Layout = (props: any) => {
   return (
     <div className={`container ${classes.layout}`}>
-      <Recipes />
+      {props.children}
     </div>
   );
 };
